+++ platformio.ini
@@ -14,6 +14,7 @@ default_envs = teensy
 [env:testing]
 platform = native
 test_framework = unity
+build_type = test
 build_flags = -D NATIVE_TC -D ARDUINO=100
 build_src_filter = -<main.cpp>
 debug_test = *
